Switching to ODS generated attribute
