+++ lib/Dialect/MSFT/MSFTAttributes.cpp
@@ -16,61 +16,12 @@
 #include "mlir/IR/BuiltinTypes.h"
 #include "mlir/IR/DialectImplementation.h"
 
+#include "llvm/ADT/TypeSwitch.h"
+
 using namespace circt;
 using namespace msft;
 
-namespace circt {
-namespace msft {
-namespace detail {
-
-struct PhysLocationAttrStorage : public mlir::AttributeStorage {
-  PhysLocationAttrStorage(DeviceTypeAttr type, uint64_t x, uint64_t y,
-                          uint64_t num, StringRef entity)
-      : type(type), x(x), y(y), num(num), entity(entity) {}
-
-  using KeyTy =
-      std::tuple<DeviceTypeAttr, uint64_t, uint64_t, uint64_t, StringRef>;
-
-  bool operator==(const KeyTy &key) const {
-    return key == KeyTy(type, x, y, num, entity);
-  }
-
-  static llvm::hash_code hashKey(const KeyTy &key) {
-    return llvm::hash_combine(std::get<0>(key), std::get<1>(key),
-                              std::get<2>(key), std::get<3>(key),
-                              std::get<4>(key));
-  }
-
-  static PhysLocationAttrStorage *
-  construct(mlir::AttributeStorageAllocator &allocator, const KeyTy &key) {
-    return new (allocator.allocate<PhysLocationAttrStorage>())
-        PhysLocationAttrStorage(std::get<0>(key), std::get<1>(key),
-                                std::get<2>(key), std::get<3>(key),
-                                std::get<4>(key));
-  }
-
-  DeviceTypeAttr type;
-  uint64_t x, y, num;
-  StringRef entity;
-};
-
-} // namespace detail
-} // namespace msft
-} // namespace circt
-
-PhysLocationAttr PhysLocationAttr::get(DeviceTypeAttr type, uint64_t x,
-                                       uint64_t y, uint64_t num,
-                                       StringRef entity) {
-  return Base::get(type.getContext(), type, x, y, num, entity);
-}
-
-DeviceTypeAttr PhysLocationAttr::Type() const { return getImpl()->type; }
-uint64_t PhysLocationAttr::X() const { return getImpl()->x; }
-uint64_t PhysLocationAttr::Y() const { return getImpl()->y; }
-uint64_t PhysLocationAttr::Num() const { return getImpl()->num; }
-StringRef PhysLocationAttr::Entity() const { return getImpl()->entity; }
-
-Attribute PhysLocationAttr::parse(DialectAsmParser &p) {
+static Attribute parsePhysLocation(DialectAsmParser &p) {
   llvm::SMLoc loc = p.getCurrentLocation();
   StringRef devTypeStr;
   uint64_t x, y, num;
@@ -88,13 +39,39 @@ Attribute PhysLocationAttr::parse(DialectAsmParser &p) {
     return Attribute();
   }
   DeviceTypeAttr devTypeAttr = DeviceTypeAttr::get(ctxt, *devType);
-  auto phy = PhysLocationAttr::get(devTypeAttr, x, y, num, entity.getValue());
+  auto phy = PhysLocationAttr::get(p.getBuilder().getContext(), devTypeAttr, x,
+                                   y, num, entity.getValue());
   return phy;
 }
 
-void PhysLocationAttr::print(DialectAsmPrinter &p) const {
-  p << "physloc<" << stringifyDeviceType(Type().getValue()) << ", " << X()
-    << ", " << Y() << ", " << Num() << ", \"" << Entity() << "\">";
+static void print(DialectAsmPrinter &p, PhysLocationAttr a) {
+  p << "physloc<" << stringifyDeviceType(a.getDevType().getValue()) << ", "
+    << a.getX() << ", " << a.getY() << ", " << a.getNum() << ", \""
+    << a.getEntity() << "\">";
 }
 
-void MSFTDialect::registerAttributes() { addAttributes<PhysLocationAttr>(); }
+Attribute MSFTDialect::parseAttribute(DialectAsmParser &p, Type type) const {
+  StringRef attrName;
+  if (p.parseKeyword(&attrName))
+    return Attribute();
+  if (attrName == "physloc")
+    return parsePhysLocation(p);
+  p.emitError(p.getNameLoc(), "Unexpected msft attribute '" + attrName + "'");
+  return Attribute();
+}
+
+void MSFTDialect::printAttribute(Attribute attr, DialectAsmPrinter &p) const {
+  TypeSwitch<Attribute>(attr)
+      .Case([&p](PhysLocationAttr a) { print(p, a); })
+      .Default([](Attribute) { llvm_unreachable("Unexpected attribute"); });
+}
+
+void MSFTDialect::registerAttributes() {
+  addAttributes<
+#define GET_ATTRDEF_LIST
+#include "circt/Dialect/MSFT/MSFTAttributes.cpp.inc"
+      >();
+}
+
+#define GET_ATTRDEF_CLASSES
+#include "circt/Dialect/MSFT/MSFTAttributes.cpp.inc"
